pyproject.toml
[tool.poetry]
name = "traveltaf"
version = "0.1.0"
description = ""
authors = ["esaatechnology"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
django-browser-reload = "^1.17.0"
gunicorn = "^23.0.0"
django-cors-headers = "^4.6.0"
whitenoise = "^6.8.2"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
requests = "^2.31.0"
google-cloud-logging = "^3.8.0"
social-auth-app-django = "^5.4.2"
openai = "^1.57.0"
pdfplumber = "^0.11.4"
load-dotenv = "^0.1.0"
mysql-connector-python = "^9.1.0"
psycopg2-binary = "^2.9.10"
django-ckeditor-5 = "^0.2.15"
django-storages = "^1.14.4"
google-cloud-storage = "^2.19.0"
django-payments = "^3.0.1"
stripe = "^11.3.0"
aiohttp = "^3.11.11"
jinja2 = "^3.1.5"
markdown2 = "^2.5.3"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
